# src/main/java/com/jiraira/prueba/infrastructure/adapters/input/rest/data/response/ErrorResponse.java
package com.jiraira.prueba.infrastructure.adapters.input.rest.data.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ErrorResponse {
    private int status;
    private String message;
    private Object data;
}